Tidy up number types

+++ src/mix/address.rs
@@ -1,4 +1,4 @@
-use super::{Sign};
+use super::Sign;
 
 #[derive(Debug, Copy, Clone)]
 pub struct Address {
@@ -7,7 +7,7 @@ pub struct Address {
 }
 
 impl Address {
-  pub fn new(value: i64) -> Address {
+  pub fn new(value: isize) -> Address {
     let sign = if value.signum() < 0 {
       Sign::Negative
     } else {
@@ -22,7 +22,7 @@ impl Address {
   }
 
   pub fn value(&self) -> isize {
-    let magnitude = ((self.bytes[0] as u16) * 64 + (self.bytes[1] as u16)) as isize;
+    let magnitude = (self.bytes[0] as isize) * 64 + (self.bytes[1] as isize);
 
     let sign = if self.sign == Sign::Positive { 1 } else { -1 };
     magnitude * sign
